+++ cmd/b3scalectl/cli.go
@@ -2,56 +2,20 @@ package main
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
-	"net/url"
 	"os"
 	"strings"
 	"time"
 
 	"github.com/urfave/cli/v2"
 
-	"github.com/b3scale/b3scale/pkg/bbb"
 	"github.com/b3scale/b3scale/pkg/config"
 	"github.com/b3scale/b3scale/pkg/http/api"
-	"github.com/b3scale/b3scale/pkg/store"
 )
 
 // RetNoChange indicates the return code, that no
 // change was applied.
 const RetNoChange = 64
-
-// Frontend retrieval helper
-func getFrontendByKey(
-	ctx context.Context, c api.Client, key string,
-) (*store.FrontendState, error) {
-	frontends, err := c.FrontendsList(ctx, url.Values{
-		"key": []string{key},
-	})
-	if err != nil {
-		return nil, err
-	}
-	if len(frontends) > 0 {
-		return frontends[0], nil
-	}
-	return nil, nil
-}
-
-// Backend retrieval helper
-func getBackendByHost(
-	ctx context.Context, c api.Client, key string,
-) (*store.BackendState, error) {
-	backends, err := c.BackendsList(ctx, url.Values{
-		"host": []string{key},
-	})
-	if err != nil {
-		return nil, err
-	}
-	if len(backends) > 0 {
-		return backends[0], nil
-	}
-	return nil, nil
-}
 
 // Cli is the main command line interface application
 type Cli struct {
@@ -344,483 +308,6 @@ func (c *Cli) showStatus(ctx *cli.Context) error {
 	fmt.Println("   api version:", status.API)
 	fmt.Println("")
 
-	return nil
-}
-
-// setFrontend manages a forntend
-func (c *Cli) setFrontend(ctx *cli.Context) error {
-	dry := ctx.Bool("dry")
-
-	// Args should be frontend key
-	if ctx.NArg() < 1 {
-		return fmt.Errorf("require: <frontend key>")
-	}
-
-	key := ctx.Args().Get(0)
-	secret := ctx.String("secret")
-
-	fmt.Println("getting frontend:", key)
-
-	client, err := apiClient(ctx)
-	if err != nil {
-		return err
-	}
-
-	// Get or create frontend
-	state, err := getFrontendByKey(ctx.Context, client, key)
-	if err != nil {
-		return err
-	}
-
-	if state == nil {
-		fmt.Println("creating frontend")
-		// Create frontend
-		if secret == "" {
-			return fmt.Errorf("secret may not be empty")
-		}
-		state = store.InitFrontendState(&store.FrontendState{
-			Frontend: &bbb.Frontend{
-				Key:    key,
-				Secret: secret,
-			},
-		})
-		if ctx.IsSet("opts") {
-			if err := json.Unmarshal(
-				[]byte(ctx.String("opts")), &state.Settings); err != nil {
-				return err
-			}
-		}
-		if !dry {
-			state, err = client.FrontendCreate(ctx.Context, state)
-			if err != nil {
-				return err
-			}
-			fmt.Println("created frontend:", state.ID, state.Frontend.Key)
-		} else {
-			fmt.Println("skipped creating frontend")
-		}
-	} else {
-		// Update Frontend
-		changes := false
-		if ctx.IsSet("secret") {
-			if secret == "" {
-				return fmt.Errorf("secret may not be empty for update")
-			}
-			changes = true
-			state.Frontend.Secret = secret
-		}
-
-		if ctx.IsSet("opts") {
-			changes = true
-		}
-
-		if !changes {
-			fmt.Println("no changes")
-			c.returnCode = RetNoChange
-		} else {
-			if !dry {
-				if ctx.IsSet("opts") {
-					// Update frontend settings using raw payload to
-					// convey explicit null values.
-					payload, err := json.Marshal(map[string]json.RawMessage{
-						"settings": []byte(ctx.String("opts")),
-					})
-					if err != nil {
-						return err
-					}
-					_, err = client.FrontendUpdateRaw(ctx.Context, state.ID, payload)
-					if err != nil {
-						return err
-					}
-				} else {
-					_, err := client.FrontendUpdate(ctx.Context, state)
-					if err != nil {
-						return err
-					}
-				}
-				fmt.Println("updated frontend")
-			} else {
-				fmt.Println("skipped saving changes in frontend")
-			}
-		}
-	}
-
-	return nil
-}
-
-// deleteFrontend removes a frontend from the cluster
-func (c *Cli) deleteFrontend(ctx *cli.Context) error {
-	dry := ctx.Bool("dry")
-
-	// Args should be host
-	key := ctx.Args().Get(0)
-	if key == "" {
-		return fmt.Errorf("need frontend key for delete")
-	}
-
-	client, err := apiClient(ctx)
-	if err != nil {
-		return err
-	}
-	state, err := getFrontendByKey(ctx.Context, client, key)
-	if err != nil {
-		return err
-	}
-	if state == nil {
-		return fmt.Errorf("no such frontend")
-	}
-
-	if dry {
-		fmt.Println("skipping delete (dry)")
-		return nil
-	}
-
-	fmt.Println("delete frontend:", state.ID)
-	_, err = client.FrontendDelete(ctx.Context, state)
-	return err
-}
-
-// showFrontend displays information about a frontend
-func (c *Cli) showFrontend(ctx *cli.Context) error {
-	key := ctx.Args().Get(0)
-	if key == "" {
-		return fmt.Errorf("need frontend key for showing info")
-	}
-
-	client, err := apiClient(ctx)
-	if err != nil {
-		return err
-	}
-
-	state, err := getFrontendByKey(ctx.Context, client, key)
-	if err != nil {
-		return err
-	}
-	if state == nil {
-		return fmt.Errorf("no such frontend")
-	}
-
-	fmt.Println("Frontend:", state.Frontend.Key)
-	fmt.Println("Settings:")
-	s, _ := json.MarshalIndent(state.Settings, "   ", " ")
-	fmt.Println(string(s))
-	return nil
-}
-
-// show a list of all frontends
-func (c *Cli) showFrontends(ctx *cli.Context) error {
-	client, err := apiClient(ctx)
-	if err != nil {
-		return err
-	}
-	frontends, err := client.FrontendsList(ctx.Context, nil)
-	if err != nil {
-		return err
-	}
-
-	for _, f := range frontends {
-		settings, _ := json.Marshal(f.Settings)
-		fmt.Printf("%s\t%s\t%s\t%s\n", f.ID, f.Frontend.Key, f.Frontend.Secret, settings)
-	}
-	return nil
-}
-
-// setBackend manages the backends in the cluster
-func (c *Cli) setBackend(ctx *cli.Context) error {
-	adminState := ctx.String("state")
-	dry := ctx.Bool("dry")
-
-	// Args should be host
-	host := ctx.Args().Get(0)
-	if !strings.HasPrefix(host, "http") {
-		return fmt.Errorf("host should start with http(s)://")
-	}
-	if !strings.HasSuffix(host, "/") {
-		host += "/"
-	}
-
-	client, err := apiClient(ctx)
-	if err != nil {
-		return err
-	}
-
-	// Check if backend exists
-	state, err := getBackendByHost(ctx.Context, client, host)
-	if err != nil {
-		return err
-	}
-	if state == nil {
-		if !ctx.IsSet("secret") {
-			return fmt.Errorf("need secret to create host")
-		}
-		// Create Backend
-		state = store.InitBackendState(&store.BackendState{
-			Backend: &bbb.Backend{
-				Host:   host,
-				Secret: ctx.String("secret"),
-			},
-			AdminState: adminState,
-		})
-		if ctx.IsSet("opts") {
-			if err := json.Unmarshal(
-				[]byte(ctx.String("opts")), &state.Settings); err != nil {
-				return err
-			}
-		}
-		if !dry {
-			state, err = client.BackendCreate(ctx.Context, state)
-			if err != nil {
-				return err
-			}
-			fmt.Println("created backend:", state.ID, state.Backend.Host)
-		} else {
-			fmt.Println("skipped creating backend")
-		}
-	} else {
-		// The state is known to use. Just make updates
-		changes := false
-		if ctx.IsSet("secret") {
-			secret := ctx.String("secret")
-			if secret == "" {
-				return fmt.Errorf("secret may not be empty")
-			}
-			if state.Backend.Secret != secret {
-				state.Backend.Secret = secret
-				changes = true
-			}
-		}
-		if ctx.IsSet("state") {
-			if state.AdminState != adminState {
-				state.AdminState = adminState
-			}
-			changes = true
-		}
-		if ctx.IsSet("opts") {
-			changes = true
-		}
-		if changes {
-			if !dry {
-				if ctx.IsSet("opts") {
-					// Update backend settings using raw payload to
-					// convey explicit null values.
-					payload, err := json.Marshal(map[string]json.RawMessage{
-						"settings": []byte(ctx.String("opts")),
-					})
-					if err != nil {
-						return err
-					}
-					_, err = client.BackendUpdateRaw(ctx.Context, state.ID, payload)
-					if err != nil {
-						return err
-					}
-				} else {
-					state, err = client.BackendUpdate(ctx.Context, state)
-					if err != nil {
-						return err
-					}
-				}
-				fmt.Println("updated backend")
-			} else {
-				fmt.Println("skipping backend update")
-			}
-		} else {
-			fmt.Println("no changes")
-			c.returnCode = RetNoChange
-		}
-	}
-	return nil
-}
-
-// showBackends displays information about our backend
-func (c *Cli) showBackend(ctx *cli.Context) error {
-	host := ctx.Args().Get(0)
-	if host == "" {
-		return fmt.Errorf("need host for showing backend info")
-	}
-
-	client, err := apiClient(ctx)
-	if err != nil {
-		return err
-	}
-	// Check if backend exists
-	backend, err := getBackendByHost(ctx.Context, client, host)
-	if err != nil {
-		return err
-	}
-	if backend == nil {
-		return fmt.Errorf("backend not found")
-	}
-
-	fmt.Println("Backend:", backend.Backend.Host)
-	fmt.Println("Settings:")
-	s, _ := json.MarshalIndent(backend.Settings, "  ", "  ")
-	fmt.Println(string(s))
-
-	return nil
-}
-
-// showBackends displays a list of our backends
-func (c *Cli) showBackends(ctx *cli.Context) error {
-	client, err := apiClient(ctx)
-	if err != nil {
-		return err
-	}
-	// Check if backend exists
-	backends, err := client.BackendsList(ctx.Context, nil)
-	if err != nil {
-		return err
-	}
-	for _, b := range backends {
-		ratio := 0.0
-		if b.MeetingsCount > 0 {
-			ratio = float64(b.AttendeesCount) / float64(b.MeetingsCount)
-		}
-		settings, _ := json.Marshal(b.Settings)
-		fmt.Printf("%s\n  Host:\t %s\n", b.ID, b.Backend.Host)
-		fmt.Printf("  Settings:\t %v\n", string(settings))
-		fmt.Printf("  NodeState:\t %s\t", b.NodeState)
-		fmt.Printf("  AdminState:\t %s\n", b.AdminState)
-		fmt.Printf("  MC/AC/R:\t %d/%d/%.02f\n",
-			b.MeetingsCount,
-			b.AttendeesCount,
-			ratio)
-		fmt.Printf("  LoadFactor:\t %v\n", b.LoadFactor)
-		fmt.Printf("  Latency:\t %v\n", b.Latency)
-		if b.NodeState == "error" && b.LastError != nil {
-			fmt.Println("  LastError:", *b.LastError)
-		}
-		fmt.Println("")
-	}
-
-	return nil
-}
-
-// enable a backend means setting the admin state
-// to ready
-func (c *Cli) enableBackend(ctx *cli.Context) error {
-	dry := ctx.Bool("dry")
-	// Args should be host
-	if ctx.NArg() < 1 {
-		return fmt.Errorf("require: <host>")
-	}
-	host := ctx.Args().Get(0)
-	return c.setBackendAdminState(ctx, host, dry, "ready")
-}
-
-func (c *Cli) disableBackend(ctx *cli.Context) error {
-	dry := ctx.Bool("dry")
-	// Args should be host
-	if ctx.NArg() < 1 {
-		return fmt.Errorf("require: <host>")
-	}
-	host := ctx.Args().Get(0)
-	return c.setBackendAdminState(ctx, host, dry, "stopped")
-}
-
-func (c *Cli) setBackendAdminState(
-	ctx *cli.Context,
-	host string,
-	dry bool,
-	adminState string,
-) error {
-	if !strings.HasPrefix(host, "http") {
-		return fmt.Errorf("host should start with http(s)://")
-	}
-	if !strings.HasSuffix(host, "/") {
-		host += "/"
-	}
-
-	client, err := apiClient(ctx)
-	if err != nil {
-		return err
-	}
-
-	state, err := getBackendByHost(ctx.Context, client, host)
-	if err != nil {
-		return err
-	}
-	if state == nil {
-		return fmt.Errorf("backend not found")
-	}
-	// The state is known to use. Just make updates
-	changes := false
-	if state.AdminState != adminState {
-		state.AdminState = adminState
-		changes = true
-	}
-	if changes {
-		if !dry {
-			state, err = client.BackendUpdate(ctx.Context, state)
-			if err != nil {
-				return err
-			}
-			fmt.Println("updated backend admin state")
-		} else {
-			fmt.Println("skipping backend admin state update")
-		}
-	} else {
-		fmt.Println("no changes")
-		c.returnCode = RetNoChange
-	}
-	return nil
-}
-
-// delete backend removes a backend from the store
-func (c *Cli) deleteBackend(ctx *cli.Context) error {
-
-	dry := ctx.Bool("dry")
-	force := ctx.Bool("force")
-	// Args should be host
-	if ctx.NArg() < 1 {
-		return fmt.Errorf("require: <host>")
-	}
-
-	client, err := apiClient(ctx)
-	if err != nil {
-		return err
-	}
-	host := ctx.Args().Get(0)
-	state, err := getBackendByHost(ctx.Context, client, host)
-	if err != nil {
-		return err
-	}
-	if state == nil {
-		return fmt.Errorf("no such backend")
-	}
-
-	// Check if we should hard delete. This can be either
-	// the case when the deletion is forced, or when
-	// the backend is in a non ready state
-	hardDelete := force ||
-		(state.NodeState != "ready" && state.NodeState != "init")
-
-	if hardDelete {
-		// The node is down anyhow we can issue a direct delete
-		if dry {
-			fmt.Println("skipping delete backend (dry run)")
-			return nil
-		}
-		fmt.Println("deleting backend")
-		state, err = client.BackendDelete(
-			ctx.Context, state, url.Values{
-				"force": []string{"true"},
-			})
-		if err != nil {
-			return err
-		}
-	} else {
-		// Otherwise, we mark the node for decommissioning
-		state.AdminState = "decommissioned"
-		if dry {
-			fmt.Println("skipping decommissioning backend (dry run)")
-			return nil
-		}
-		state, err = client.BackendUpdate(ctx.Context, state)
-		if err != nil {
-			return err
-		}
-	}
-	fmt.Println("backend marked for decommissioning")
 	return nil
 }
